remove Complaint from models and move utils method from dgf
