+++ openprocurement/auctions/core/awarding_2_0/migration.py
@@ -2,7 +2,7 @@ from barbecue import chef
 from uuid import uuid4
 
 from openprocurement.api.utils import get_now
-from openprocurement.auctions.dgf.utils import invalidate_bids_under_threshold
+from openprocurement.auctions.core.awarding_2_0.utils import invalidate_bids_under_threshold
 
 
 def switch_auction_to_unsuccessful(auction):
